+++ core/src/main/kotlin/com/tschuchort/compiletesting/kapt/util.kt
@@ -1,9 +1,9 @@
 package com.tschuchort.compiletesting.kapt
 
 import java.io.File
-import org.jetbrains.kotlin.kapt.cli.KaptCliOption
 import org.jetbrains.kotlin.kapt.base.KaptFlag
 import org.jetbrains.kotlin.kapt.base.KaptOptions
+import org.jetbrains.kotlin.kapt.cli.KaptCliOption
 
 fun KaptOptions.Builder.toPluginOptions(): List<String> {
     val options = mutableListOf<String>()