+++ CHANGELOG.md
@@ -4,6 +4,12 @@ Changelog
 **Unreleased**
 --------------
 
+- Remove support for Kotlin `1.x`. Only Kotlin 2.x is supported now.
+  - This includes removal of all legacy 1.x compiler APIs such as `ComponentRegistrar`.
+- Remove support for KSP1. Only KSP2 is supported now.
+- Changed `verbose` to be `false` by default.
+- Build against KSP `2.2.20-2.0.3`.
+
 0.9.0
 -----